i pulled the news feed out of the insult file and gave it a new file of its own
